# RFC: construct the name to object dicts lazily