Following up on the discussion in https://github.com/JuliaOpt/JuMP.jl/issues/1402. This change avoids creating the `names_to_var` dictionary until it's needed.

Using @IainNZ's script that benchmarks variable creation time,

Before:
```
10^1:  0.000 s, 3981.300000 ns/var, 286 a, 28.600000 a/var
10^2:  0.000 s, 1016.850000 ns/var, 844 a, 8.440000 a/var
10^3:  0.001 s, 703.424000 ns/var, 6756 a, 6.756000 a/var
10^4:  0.007 s, 664.980500 ns/var, 69774 a, 6.977400 a/var
10^5:  0.100 s, 1003.817160 ns/var, 699808 a, 6.998080 a/var
10^6:  1.558 s, 1558.397552 ns/var, 6999859 a, 6.999859 a/var
```

After:
```
10^1:  0.000 s, 3638.900000 ns/var, 286 a, 28.600000 a/var
10^2:  0.000 s, 829.310000 ns/var, 838 a, 8.380000 a/var
10^3:  0.001 s, 520.511000 ns/var, 6742 a, 6.742000 a/var
10^4:  0.005 s, 491.230300 ns/var, 69754 a, 6.975400 a/var
10^5:  0.077 s, 766.043240 ns/var, 699776 a, 6.997760 a/var
10^6:  1.005 s, 1004.825593 ns/var, 6999809 a, 6.999809 a/var
```
so we get the ~0.5 seconds back as predicted.

These results are with an immutable `VariableInfo` struct (https://github.com/JuliaOpt/JuMP.jl/issues/1525).

Closes #536